Make sure the autoupdate process is waited on.
Without this, the monitor will create a massive army of zombies during development. Not that I have anything against zombie armies, I just feel that an evil mastermind should put some effort into it.

According to sigaction(2), setting SIGCHLD to SIG_IGN should also work (in that I won't have to wait() for the child), but that's not very portable. It might be nice to run on other machines in the future. So what I do is just have a separate thread that executes waitpid. The thread of course is detached, so I don't have to wait for winter for the gorillas to freeze to death.