Database version 10: Add unique(from_id, to_id).
Adding the uniqueness to links is nice because it simplifies the link comparison logic when output links in the Tupfile change. Also, it is nice because that better signifies my intention with the links anyway. When output links change, tup will now re-run the command in order to check that the links are valid. These tests are t6024, t6025, and t6026.