Fixed a bug that caused the updater to quit early.
This apparently goes back to database version 2 - if the updater came across an unexpanded node, it would skip the check for returning workers. It is possible that plist would be empty in this case, and so it would break out of the loop while commands were still working. This would cause the fatal graph empty error. I'm not sure if this is the same error that I came across while building the kernel, but it's highly plausible. Anyway, I added some code to at least display what was left in the graph to help debugging.