Use correct pattern for symlinks.
The problem with using 'ln -sf %f %o' is that if the symlink already exists and points to a ghost node, if the command runs again for whatever reason, then ln will try to stat the file that doesn't exist and get ENOENT. This causes ldpreload to report a ghost-read on the destination of the symlink. The reason this is bad is because it can randomly break (ie: if the file being made a symlink to is generated, then order matters).

Another option would be to assume a symlink always implies a read on the 'from' file, but that would needlessly re-create the symlink everytime the 'from' file changes.