Update todo.
I decided it would be a bit more work to move start_server() out since I rely on the fact that stop_server() joins the pthread in order to make sure the message queue is flushed. If I keep the pthread there and just move the socketpair out, I don't gain a whole lot, and it adds more functions and confuses things.