More efficient reclaim_ghosts query.
For some reason with the 'left join on dir=ghost_list.id or sym=ghost_list.id', it won't use the two indexes (one on dir and one on sym). Instead it iterates through the whole node table, which naturally sucks performance-wise. The way it's written now it goes in effectively two stages - one on dir/link (which can use both indexes), and then afterward on sym (which now can also use the index).