Added t2055 - .gitignore functionality.
Since tup already knows the files that are generated and thus don't belong in revision control, it seems reasonable to generate the .gitignore file. I'm not sure if marking the .gitignore file as TUP_NODE_GENERATED without a command to officially generate it will cause problems. Also I haven't tested it significantly with git itself, so it may behave in bizarre ways. But, it seems like a good first step, and is only enabled with .gitignore in the Tupfile.