Added t2057, t2058 - more gitignore stuff.
Now you can specify .gitignore anywhere in the Tupfile and it will only create it at the end of parsing (to make sure all rules are processed by that point). Also I make sure that the top-level .gitignore is created if requested, even if there are no output files in that directory.