Added t7026 - .gitignore/monitor special case.
The problem is that watch_path ignores all files that begin with '.', but the .gitignore node is put into the database by the parser in order to keep track of it (in case the .gitignore directive is removed from the Tupfile later). The monitor would then move the node to delete, and watch_path would skip it, so the directory would end up getting put back into create. Now I made it so all .gitignore files don't go into the tmplist for checking, since it seems to be the easiest way.