Added t2059 - make sure dependent directories are parsed.
Had a similar issue with the updater when a failed node wasn't re-executed since it wasn't put in modify. Since the create phase is only committed if they all succeed, this currently isn't an issue. However, it is nice to have the test since if I change the create phase in the future, this could break.