Added t7027 - correctly handle files moved away from tup.
The monitor was not properly handling the case where there is an IN_MOVED_FROM event without a corresponding IN_MOVED_TO event. This case needs to be handled the same as a regular deletion. To handle this, instead of keeping a separate list of moved_from_events, the moved_from_list now just has a list of pointers to events in the queue. Any moved_from_events that are claimed by IN_MOVED_TO events are removed from the list, and the flags are cleared so they won't be processed by handle_event(). Any that aren't claimed will be deleted and then removed from the list.