Add sql debugging mode.
Now you can add a --debug-sql flag (only in the first argument, because I'm lazy), and tup will print the SQL statements as it executes them. It's not quite SQL format because it re-uses the sql string (which has '?' instead of '%i', for example). This is so the debug will stay in sync with the actual code. Also, I'm lazy.

Hopefully it will be easier to track down inefficiencies or other incorrect behavior.