Added t5048 - directory moving test with dependency.
This is actually a use-case to test the need for tup_db_modify_deleted_deps(). When I tried commenting out the function call in process_delete_nodes(), all tests still passed. I assume this is because that functionality is largely already handled by tup_del_id(). In fact, it may be necessary to move all deleted nodes to be processed through tup_del_id() in order to catch all the cases.