Major parsing optimization.
Thanks to the handy --debug-sql option, I could see that I was querying and updating the link table far too often when no changes were required. Now instead of a bunch of SQL that would run twice on each link (!) (thus iterating through all the links for a node every time), I select all of the links for a node out once and build a tree in memory. Then I build another tree in memory corresponding to what the links should look like based on the Tupfile, then use a merge algorithm to determine the differences between the two. Based on the differences, the minimal amount of database changes can be applied.

The number of SQL queries in order to parse 56 Tupfiles in my current gittup tree went from 287650 to 16876, and the parsing time went from 10.5s to 1.8s.