Added t6028 - reordering commands.
This is similar to t6021, only the removed sticky link is in going to be deleted because the commands were re-ordered. This should cause the command to be re-executed so we can check if the link is still necessary.

Also added the ability to check the reason for update failures, so I can make sure the tests are failing properly.