Added t6029 - new command creates the same output file.
It shouldn't be possible for a new command to create the same output file as an old command that is being re-parsed (and hence marked for deletion). All I did was delete the old link when creating the new link, so that when the old rule is parsed again it will find out that it's no longer the owner of the file.