Use compare_trees for actual input files.
Similar to the parser optimization, save the actual input files in a tree, and compare to the tree of input links in the database before writing out the changes. The time difference isn't quite as pronounced, but still noticeable. Apparently I have some other errant queries which need to be addressed as well.