Use final pel pointer instead of final char* pointer.
The problem is if I had a path like 'foo/bar/.', or 'foo/bar/..', then the "last" pointer would point at "bar/." or "foo/bar/..", since there was no len attached. In theory I could just overwrite the last byte with a nul, but then I'd have to de-constify everything, and I didn't feel like looking into what side effects that would have. This is slightly more correct, even if it comes with the annoyance of having to duplicate a few string routines that don't take a length (like strcspn).

The end result is that I don't query for a path with a slash in it, since that won't ever exist in the database.