Added t6031 - generated file becomes a normal file.
I figured this would be useful with gittup if I wanted to move a generated file into the git repo without moving it around. Without this capability, the newly checked-out file would get deleted by tup since the rule would be removed, which would be annoying. Now if a generated node is modified and the rule is removed before an update occurs, the node will become a normal file.