Monitor now restarts itself if it gets out of whack.
Though it's kinda annoying, this is probably the only reasonable solution given that for a sufficient volume of filesystem changes, the database simply can't keep up. For example, removing a file is very quick for the filesystem, but quite laborious for tup (possibly a case for optimization). Doing things such as 'rm -rf large_tree' will cause events to get dropped, so the monitor gets out of sync. When this happens, the monitor will clear out its dircache and inotify queue, then re-read the tree from the top.