Added t3015 - bang macros.
It's kinda messy, but at least the initial go at it works. The bang macro can add in order-only prerequisites, but not to the regular input list (so it doesn't affect %f). Also, only one of the macro and the rule can have an output specified. Currently bang rules can't be chained - will see if that becomes useful later.

There are probably many errors with the parsing implementation of this :)