Added t6033, t6034 - circular dependency issues.
I fixed the introduction of the circular dependency (see t6033 for an example) by having the parser check for duplicate commands in a Tupfile. I had previously tried to also check for circular dependencies with some extra logic when building the graph, but I've since determined that it is not possible. See the comment in t6034 for a fuller explanation. Basically I have to choose between either: 1) always expanding every node when building the graph, or 2) deferring until graph execution to find all circular dependencies. I am currently going with 2, since that is already implemented, still catches all circular dependencies, and is faster in that it doesn't load parts of the DAG that it doesn't need.