Added t5054 - check removed normal links for ghosts.
I didn't see this before since I normally changed the command string when changing ghosts. Now I keep the command the same, but change the shell script so it no longer uses the ghost. When the link goes away, the ghost would stay around. Now I add the link source to the ghost list if necessary.