Added t2066 - .EMPTY bang rules.
The .EMPTY bang rule can be triggered when an input is specified (ie: like {objs} or $(obj-y) or something), but it evaluates to empty. In this case you may want a separate command to execute. For example, in the kernel Makefiles it would create built-in.o with ld -r if there are objects, but create a built-in.o which is really an empty archive if there are no objects. Then the parent Makefile could always include the subdirectory's built-in.o. This practice is now copied in the Tupfiles.