Change variables to be more compatible with kconfig.
Now I keep track of the special "tup.config" file at the tup root. Whenever it changes or gets deleted, I will read it in at the next update and change the internal @-variables to reflect the definitions in tup.config. This way kconfig doesn't have to support tup directly, and the config can be checked into git, and hopefully have it work with bisection.

Some other fixes are in here too, but I'm too lazy to separate them out.