Added t6035 - incorrect generated -> file conversion.
When running with parallel compilation, it was possible for one command to succeed and mark its output as modified (in order to continue there after errors), and another command to fail. If these commands are then removed from the system (by changing a config variable, for example), then the command that succeeded will have left its output in modify, causing it to stay around because of t6031. However, the file really should be removed.

I tried messing around with mtimes, and was even considering adding a new flag table in order to keep track of actual modifications by the user vs. flagging a modification for the updater. Fortunately the fix is much simpler, which is for commands to simply mark their output files' outputs as modified (ie: the next level of commands). This avoids marking files as modified that weren't actually modified, and still maintains the pick-up-where-you-left-off capability when errors occur.