Added t4019 - make sure "%f;" doesn't duplicate ";"
This was leftover from ye olden days, when I couldn't easily get a list of object files, so I did something like "%F.o" in the linker rule, and expected it to take "foo.c bar.c" and make it into "foo.o bar.o". That isn't necessary anymore, and is confusing if you happen to put non-space characters after the %[fFbB] variable.