Allow reading from .hidden files in commands.
Previously I was giving an error if a command read from a file hidden with a ".", but now I just don't track the dependency. This allows commands like 'git describe' to be executed in Tupfiles, but of course any files it reads in the .git directory won't be tracked. However, it can still be useful in the final linking stage, such as how tup itself builds in the version string.