Execvp behavior change, some warning fixes.
I built it on an ubuntu host and got some different warnings, which I fixed. Also the gcc there tried to execute 'as' with execvp, so I figured that might be a common occurance, and changed the execvp to allow it to pass through if it is going to use the path searching.