Put monitor pid in a regular file, not database.
This way 'tup stop' can always act instantly, rather than waiting for the monitor to get around to checking the monitor lock file. Also, I can now easily check the pid file from the updater to see if the monitor is running - if not, 'tup scan' runs automatically. This way tup always properly updates when you do 'tup upd'. Annoyingly, this broke a bunch of test cases which try to do things behind tup's back in order to check that things are updating for the right reasons. I added a --no-scan option to support that.