Replace existing commands when command strings change.
There would be a lot of database churn if there are a lot of commands that are all generated from the same or similar rules (ie: CFLAGS set in a high-level Tuprules.tup), and can be exacerbated by the large number of links of the commands. For example, changing the CFLAGS from -O2 to -Os would end up deleting all the -O2 commands, removing all their links and nodes, then adding -Os commands, which then get executed and have all their links added back again.

Now I try to use the outputs as a unique identifier of a specific command, so changing the command string will only result in re-naming the existing command in place. This saves time in both the parsing and updating stages, since the parser doesn't have to delete all those nodes, and the updater doesn't have to re-create all the links. It also means if you try to fix a failing command by changing the command string, it should run again immediately (since it has the same tupid), rather than getting stuck at the end of the queue.