Add t6036 - test missing generated files during scan.
The problem was that the generated file wasn't created yet (since the update never fully succeeded), and during a scan it would notice that the file was missing and put the directory back in the create list. This was unnecessary. Now I make the distinction between just noticing a file is missing vs. actively removing the node from the updater. In the first case, a missing generated file just puts the command back in the modify list.