Add tup_entry caching structure.
This is a full node cache of the database. It should let me issue fewer database queries, at the expense of the extra code here and there to keep it up to date. It will eventually replace the dirtree structure, and probably struct db_node as well. Initial results are promising, but right now it's very confusing with both dirtree, tup_entry, and the old-style direct db access stuff all in there at once.