Added t6037 - make sure cache is updated on name change.
I missed one of the node-changing queries, so the cache was out of date wrt the actual database. This was missed in all the test cases because it only shows up if the command is loaded before parsing, which only happens if a command is dependent on a directory. I think the only command in all of gittup.org where this is true is when creating initrd using cpio (so, that's where I found it). Anyway, the new test case mimics this behavior, and the cache is correctly updated now. Also got rid of tup_db_select_dirname(), since it is now unnecessary.