Fix invalid pointer use.
The problem was that the tup_entry could be removed before the tup_tree_entry, so the tent pointer was invalid. This can happen if the directory is removed before a subfile, since the directory recursively deletes the nodes under it. Now I just use a regular tupid_tree to track which files are still there and get the tup_entry as necessary.

Also made the tup.config get a tupid, which means the last check in t3000 is no longer valid (since the tup.config file will be in modify). I felt it was kinda confusing to try to do 'tup g tup.config' and have it give an error.