Check IN_MOVE_SELF and IN_DELETE_SELF events.
Apparently when you move a directory, inotify still tracks it for you. In this case I actually prefer not to track it (since I will add a new watch if necessary). Also it was annoying if you moved a directory outside of tup, since it would still be watched, but the tupid referenced in the dircache was no longer correct. Now I remove the watch on these two events, and then the IN_IGNORED event takes care of cleaning up the dircache.