Added t6038 - make sure a symlink to '.' works.
When the filename was '.', the pel pointer in find_dir_tupid_dt wasn't getting set, so I was trying to access a bad pointer. I'm starting to try to move things toward using tup_entry, and not have so much database logic in things like update_symlink_fileat, which is why I made a new function here. More work needs to be done to do all that though, but this at least fixes the segfault.