Added t5060 - make sure an output sym can become normal.
If an output file went from a symlink to a normal file, the sym field stayed around. Also, just adding tup_db_set_sym(-1) wouldn't work because if the symlink already existed, then the filesystem would try to write through the symlink. Now that I remove all outputs before running a command, that is no longer an issue. So whenever a normal file is written, I can always clear the sym field to -1.