Fix some error messages.
In one case, I shouldn't call tup_entry_get if the tupid is less than 0. In the other, apparently I make use of the fact that find_dir_tupid_dt returns -1 without printing an error message in some cases. I added the debug because it was annoying when that case gets triggered when you don't expect it to (which can happen if you're fiddling with code, otherwise it probably doesn't matter).