Remove superfluous tup_entry_adds() in entry.c
Well almost superfluous - I had to add the VAR_DT node to the entry cache when reading in vars, since I already know the value of VAR_DT ahead of time (unlike other directories, which have to be read from the database, and are therefore cached). Originally the tup_entry_add() calls were here only because not everything used the cache, so sometimes a directory wouldn't be there when a subdirectory or file would try to be added.