'tup init' checks for an already initialized db.
I used to only check if a .tup/db file was already present. Now I actually go up the tree and check if a .tup/db file is anywhere in the tree. Since this screws me on how I normally test, I added a --force flag that will create a .tup directory even if one exists up the tree.