Fixed t3019 - shouldn't 'tup rm tup.config'
By telling tup that I removed tup.config, it went through a different path than the scanner. If the scanner found that tup.config was removed, it wouldn't try to re-parse the variables, so it would leave them all in the DAG. This is obviously wrong. To fix it, I make both tup_file_del and tup_file_missing check for the removed config. I still don't check in tup_del_id_force, since that only is called when removing a directory or removing a generated file. This doesn't affect tup.config, since you can't remove "." and still have a valid tup db, and tup.config can't be generated.