Writing outputs from the parser now done in memory.
I'm trying to get rid of the tmp_list, and the parser used it to store a temporary list of outputs for a command. Then I used some wacky queries on the tmp_list in order to get the diff into the database. Now I just read out the outputs into a list and compare it to the tree of output ids like I do for inputs. It didn't speed up things quite like I hoped, but it is slightly faster. I think most of the bottleneck in the parser now is from the fact that I delete and re-create the directory-level DAG everytime.