Bugfix with ghost variables.
I don't have a good testcase for this, since it happens when a user-program (like gittup.org's gcc) accesses a variable in order to make a ghost variable. The problem was that when I access a ghost variable in a Tupfile, I added an entry for it in the var table as well as the node table. However, when it came from a user program, it would only go in the node table. This changed the behavior of get_db_var_tree(), since it wouldn't get the ghost variable from the database (which does a join on node, var). So instead of going through compare_vars() which had a check for ghost nodes, it would go through add_var(), which went to tup_db_create_node_part, which would change the type to TUP_NODE_VAR, but not set any flags.

Now I changed the behavior so a ghost variable never has an entry in var, and made tup_db_create_node_part add the flags as necessary. The checks for ghost nodes in remove_var and compare_vars are no longer necessary, since get_db_var_tree() won't get the ghost variables from the database anymore.

I only noticed this because I was fiddling with the OSS_AUDIO variable in mplayer, which isn't used in any Tupfiles, but is used in .c files.