'tup init' bugfix.
So apparently all that fuss I went through to make sure 'tup init' wouldn't run if there was already a .tup directory up the tree caused 'tup init' to break when there wasn't. Not sure how that happened. Anyway, now it will fchdir back to the starting directory before trying to create .tup/ rather than trying to create it at the root of the filesystem.