Optimize @-variable db queries.
I was doing two queries each time a variable was used in the Tupfile - now I just do one and cache it for future use. A further optimization can be done when reading the @-variables from the file, but that adds some extra complications that are more difficult to verify.