Added t7034: Ephemeral directory test.
I don't want to remove directory events like I do file events, so they should never pass the ephemeral test.

Also added error checking to handle_event, so it is easy to see in the monitor tests if the monitor quit prematurely due to an error. The stop_monitor function will display an error message in such a case.