Added t7035 - check inotify watches on dir moves.
When a directory was moved outside of tup, the subdirectory inotify watches would still be active. This led to mass confusion. Now when a dir is removed in this fashion, I use a callback in order to tell the monitor to remove all watches right away. I removed the use of IN_IGNORED events for this, since those events seem to be delayed, even after flushing the monitor. I also no longer need to catch the IN_MOVE_SELF and IN_DELETE_SELF events, since I will be watching the parent directory in all cases.