Added t7036 - monitor broke when moving directories.
The problem was if ghosts were in the directory, then the directory itself would become a ghost. When this ghost was reclaimed, the corresponding tup_entry in the monitor wasn't removed. To fix this, I made reclaim_ghosts() go through a list of ghosts one by one and call delete_node() which will handle the tup_entrys properly.

This has the pleasant side effect of removing ghost_list from the database, and just using an in-memory list instead. A bunch of queries go away, and the functions that are left are much simpler.

Also added t5062, since I screwed up the debug ghost_check command when re-organizing things and hosed up my database :)