Added t3024 - allow 'foreach' in !-macros.
Now you can specify foreach when defining a !-macro, so when the macro is used it will always do a foreach. For example, this could be applied to a !cc macro so it will always compile all C files individually. There is no way to cancel a foreach in the !-macro.

Also I changed the behavior of !-macro aliasing. Now it makes a deep copy, so if you override the original macro, the aliased one won't change along with it. This was because I added the 'foreach' flag, which is specific to each bang_rule structure. So if you changed a macro from a foreach to a non-foreach, the flag would be wrong in the aliased version.