Simplify some db queries.
When parsing a directory, I mark all of the commands and generated nodes for deletion in case they were removed. I was doing this by marking everything that is TUP_NODE_CMD, as well as everything linked to by a TUP_NODE_CMD. This is done more simply by all TUP_NODE_CMD and TUP_NODE_GENERATED. Also since the .gitignore is TUP_NODE_GENERATED, it is no longer a special case.