Change t3026/t3027 - make *-chain variables.
It was pretty confusing trying to shoe-horn in chains of !-macros in both the forward and reverse rules, as well as in aliasing the !-macros themselves. So instead I created a new kind of variable, which is a chain of !-macros, called a *-chain. (I'm running out of special characters to use). Now a :-rule can only use one !-macro or *-chain, rather than specify multiple !-macros directly.