Added t5064, t5065 - mkstemp/mkostemp support.
Some programs (such as ccache) write to the output files by generating a temporary file using 'mkstemp', and then renaming that file to the output. This is similar to how 'ar' works, but apparently 'ar' doesn't use mkstemp to generate the temporary file. Now mkstemp/mkostemp files are treated as if they were opened for write.