Fix 64-bit warnings.
I should be passing an int* through the pthread_create, rather than assuming an int and void* are the same size. Also I'm not sure of the proper way to use "%.*s" with a sizeof() value for the length, so right now I'm just casting it to an int since that's what it expects.