Added t6041 - circular dependency fix.
Tup was creating a circular dependency in the DAG if a Tupfile claimed a command would write to a file, but the command instead only read from that file. Since the input files were only pruned if they matched a file that was actually written, the file would end up with a link both from and to the command node. A future update would then fail with a circular dependency.

The fix here is to skip any input tupids that match a tupid in the output tree pulled from the database.