Remove timestamp updating of 'tup touch'
All I really care about when I do 'tup touch' is that: a) it creates a file if none exists, and b) it marks the node as modify. Trying to actually update the timestamp in the filesystem is pretty useless since 'tup touch' is only used for testing, and I don't really need to preserve the semantics of just 'touch'. As far as I can tell, this was never actually necessary.