Added t6042 - have a file become a directory and vice versa.
Don't know how I didn't stumble across this before. If a node exists as a file, and then you remove it and create it as a directory (without the monitor running), then during the scan tup gets confused and quits. Instead it should assume that the old file/directory was removed and create a new one of the appropriate type.