Change ldpreload socket to be a stream.
Apparently relying on the DGRAM socketpair to work reliably across multiple processes is a bad idea, since it only works on linux and solaris. On BSD variants it doesn't work at all - once the socket is full, it just returns errors instead of blocking. Also I can't lock the socket descriptor itself (at least, not on BSD...again that only works on linux and solaris), so I have to make lock files, one for each job slot.

I made the job slots directories. In the future I may redirect all written files to these directories instead of going straight to the filesystem for additional reliability guarantees. For example, a job may only be able to unlink() a file that exists in its job slot, not just any file in the filesystem.