libtup_client uses tup_send_event from ldpreload.
In order to properly use the same mutex and file lock that ldpreload is using to control the socket, libtup_client now goes through the same function. It can't call the function directly since it isn't actually linked to tup-ldpreload.so. Instead it just assumes it is preloaded and uses dlsym to get the function pointer. This way if one thread or forked process tries to access a variable through the client api while another thread/process opens a file (or whatever), they won't stomp on each other when trying to send events through the socket.