Fix segfault with 'tup g' and symlink loops.
The tup_entry_add() function will add an entry for the symlink when a node is added. If the symlink points back to the original node, this resulted in an endless loop because the first node was never put into the entry tree. Now I've moved things around so it will get put into the tree before the parent and symlink attempt to resolve.