Fix recursive-grep-killing-monitor issue.
The monitor should only be running once for a particular tup database. In order to make sure the user wouldn't inadvertently run it twice, the monitor would watch the .tup/monitor file. When the second monitor process tried to open this file, the first one would quit. Since then, the stop_monitor function was changed to instead just track the monitor pid and signal it directly to stop (rather than use inotify as a poor-man's signal). Since any other process opening the .tup/monitor file would kill the monitor (such as a recursive-grep), it was annoying. Now the second monitor process will just call stop_monitor to kill the previous monitor if necessary.