Added t5066, t5067, t6043 - failed command checks.
If a command started to write some files and then failed (either with a return code or via a signal), tup would not check the file accesses. This means a failed command could write out a bunch of files that require the user to manually clean them up. Additionally, if a command writes its output file without ever executing successfully and is then removed from the Tupfile, the output file would get resurrected (or whaver I called that when it gets printed in white). The file should actually be removed in this case - the problem is that without the file check, the mtime in the database would never be updated.