Clear the tup_entry cache if valgrind is running.
This is determined if the TUP_VALGRIND environment variable is set. Normally the cache is persistent until the tup program quits, so the OS can just clean it up. Unfortunately that looks like a leak to valgrind, so it would be nice for tracking down real errors if these warnings didn't clutter the output.