Added t5069, t5070 - allow transitive dependencies.
Now when a missing input dependency is found, we will try to build a connecting graph from the missing input to the command that failed to specify it. This is useful in situations where a shared library is dependent on other shared libraries that a program may not be interested in knowing about.

This does, however, have two negative side effects:
 - We have to re-build the graph from the input to the command, which is
 only bounded by the partial DAG of the input. So, performance may be affected if this case is triggered a lot.
 - In order to make sure the inherited dependencies are consistent, it
 is imperative that the command inheriting dependencies is re-executed in case the command it is inheriting from no longer uses that input. This is why the test uses a "strongly connected" graph, which doesn't allow sticky-only links to be used. As a result, this may mean that transitive dependecies precludes the use of short-circuiting builds if outputs are unchanged.