Prevent setting CONFIG_* variables.
Since $(CONFIG_FOO) is an alias for @(FOO), it won't do any good to try to set CONFIG_FOO in a Tupfile. The @-variables must be set in tup.config. Now you get an error message if you try to set such a variable, as opposed to silently ignoring you.