Added t6045 - circular dependency (broken update).
The test case is rather simple, but only shows up as a result of the work done a while ago to try to reduce the database churn. Basically because the command node is re-used based on the output file, there was a way to get it to add a circular dependency into the DAG. This would then cause the update to give a fatal graph-not-empty error. The solution is that when a link is demoted from normal+sticky to normal, we need to check if the source of the link is in the delete list. If so, the link needs to be removed entirely. This is safe because the command gets put in the modify list, so the link can be re-added if necessary.