Added t5071 - removed file shouldn't be resurrected.
Once a file is touched, tup will put it in the modify list. If it is subsequently removed (and it's a generated file), normally tup would just set it's command to modify in order to re-generate it. However, if this command is removed in the same update, the node should no longer be in the DAG.