Added t7037 - monitor shouldn't die over ephemeral directories.
If a directory is created and removed right away, the monitor will get the creation event and try to fstatat() the directory, then fail because it is already gone. This event, and the subsequent deletion event should just be ignored. Directories get skipped in the ephemeral event check due to t7034.