Re-work the monitor status messages again.
monitor_get_pid() now returns 0 if the monitor is not running, and -1 if there was an error determining if it's running or not. Also the 'Restarting the monitor' message that I added in d40f64eba is moved to monitor_get_pid() since we only want to display that in the one case where the pid doesn't exist anymore.