updater: Use threading primitives instead of sockets.
Previously I was relying on the OS to synchronize the main execution thread with the worker threads based on the semantics of socketpair(). However, that isn't very portable, so now I use explicit synchronization with pthreads & condition variables. Granted pthreads aren't portable to windows either, but that's hopefully easier to add a wrapper library to, and regardless should work much better with other unix-like OSes.