win32: Add fchdir implementation & open wrapper.
This differs slightly from James' branch. Instead of using a platform-agnostic fd structure, I still use int's for file descriptors, and all the mapping necessary for windows to support basic POSIX functionality is done behind the scenes. Basically an open() call on a directory is trapped and instead of going to the OS, an entry in dirpath is added with a fake file descriptor. Later when this file descriptor is used in fchdir(), the absolute path can be used to run chdir() instead.