win32: Fix pthread_cond_wait()
So it turns out I should pass in the HANDLE (not the address of the HANDLE) into WaitForSingleObject, otherwise it will fail, causing the pthread_cond_wait() loops to spin and waste tons of time. Fortunately a HANDLE is just void*, so this error was easily caught by the static type-checking system in place, and I didn't end up trying to re-implement the threading scheme three different ways. Or maybe I should learn to check return values :(