win32: Open shouldn't wrap with O_CREAT
The GetFileAttributesA() function is used to determine if it is a file or directory being opened, so directories can go through the dirpath structure. However, if O_CREAT is passed in, then it is definitely a file, and we won't be able to use GetFileAttributesA() on it since it doesn't exist yet.