win32 udp server: chdir before CreateProcess()
The working directory passed into CreateProcess() only applies after the process is created - it doesn't affect where it searches for an executable (like ./foo.exe). This normally worked anyway because the unlinkat() called to remove output files would end up chdiring to the right spot. If a command has no outputs though it can still be in a wrong directory. This fix just locks dir_mutex and calls chdir() before CreateProcess() in order to synchronize with other worker threads. Note that we can pass in NULL for the current directory to CreateProcess() since that tells it to use our cwd.