dllinject: Add dependency on the .exe and .dll
When injecting the dll, we have access to the executable name itself and the dlls that it will read in. These should be treated as ACCESS_READ dependencies. That way you can build and run a program from a Tupfile, and when the program changes, it will be re-executed. This should work if you are building a dll to link into a program as well, but that is currently untested.