udp server: Slightly more consistent error messages
Failing to create the subprocess would result in a "tup internal error" in the updater since it server_exec was returning 0 but not setting s->exited or s->signalled. Now if the process fails to get created or properly shutdown, the rc will be -1. Also the command string will be displayed by the updater, so it doesn't need to show up in the messages in udp.c