win32 open wrapper: Remove invalid error message.
There shouldn't be an error message displayed just because GetFileAttributes() failed. The failure could be expected if we are trying to open a non-existent file. In this case we should just fall through to the __real_open case (and likely get an error from the OS).

The problem is that the parser tries to open a Tupfile in a directory when it gets parsed, but doesn't bother to stat it first. It just uses the failure-to-open as an indication that tup doesn't need to do anything there, so it is confusing if that produces an error message.