Added t7038, t7039 - monitor file removal bug.
The monitor process has a cache of all tup_entry's from doing the initial search. It is possible (and likely) that a future update will remove files (ie: by removing a rule). The monitor process will still have an entry for this file. It seems sqlite can also re-use tupids for future nodes, so a later addition might have the same tupid and confuse the monitor.

Now the monitor will continue processing events while the update happens, only it will not modify the database in this mode. It will just look for file removal events in order to remove those entries from its cache in memory.