More correct bugfix from 23c42bf007 - add NULL check.
It is possible that tup_entry_find_name_in_dir() will set the tup_entry to NULL, which means we won't need to remove the entry from the cache. In that case we can just do nothing. I consider this a better option that my previous plan, which was to segfault.